+++ src/inspect_scout/__init__.py
@@ -2,6 +2,7 @@ from inspect_ai._util.deprecation import relocated_module_attribute
 
 from ._grep_scanner import grep_scanner
 from ._llm_scanner import AnswerMultiLabel, AnswerStructured, llm_scanner
+from ._project import ProjectConfig, project
 from ._recorder.recorder import (
     ScanResultsArrow,
     ScanResultsDF,
@@ -81,6 +82,8 @@ __all__ = [
     "scanjob",
     "ScanJob",
     "ScanJobConfig",
+    "project",
+    "ProjectConfig",
     "scan_list",
     "scan_status",
     "scan_results_df",