+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- [Projects](https://meridianlabs-ai.github.io/inspect_scout/projects.html) for centrally managing scanning configuration.
 - [Grep Scanner](https://meridianlabs-ai.github.io/inspect_scout/grep_scanner.html) for pattern-based scanning of transcripts.
 - Scanning: Add `--dry-run` option for previewing scanner counts.
 - Scanners: Fixup type annotations in @scanner decorator.